Add memory leak test.
