+++ python/ray/train/tests/test_torch_utils.py
@@ -8,6 +8,7 @@ from ray.air._internal.torch_utils import (
     load_torch_model,
     contains_tensor,
 )
+from ray.util.debug import _test_some_code_for_memory_leaks
 
 data_batch = pd.DataFrame({"A": [1, 2, 3], "B": [4, 5, 6]})
 
@@ -74,6 +75,24 @@ class TestConvertPandasToTorch:
             assert np.array_equal(
                 tensor.numpy(), data_batch[[data_batch.columns[i]]].to_numpy()
             )
+
+    def test_tensor_column_no_memory_leak(self):
+        # Test that converting a Pandas DataFrame containing an object-dtyped tensor
+        # column (e.g. post-casting from extension type) doesn't leak memory. Casting
+        # these tensors directly with torch.as_tensor() currently leaks memory; see
+        # https://github.com/ray-project/ray/issues/30629#issuecomment-1330954556
+        col = np.empty(1000, dtype=object)
+        col[:] = [np.ones((100, 100)) for _ in range(1000)]
+        df = pd.DataFrame({"a": col})
+        suspicious_stats = _test_some_code_for_memory_leaks(
+            desc="Testing convert_pandas_to_torch_tensor for memory leaks.",
+            init=None,
+            code=lambda: convert_pandas_to_torch_tensor(
+                df, columns=[["a"]], column_dtypes=[torch.int]
+            ),
+            repeats=10,
+        )
+        assert not suspicious_stats
 
 
 torch_module = torch.nn.Linear(1, 1)
